refactor: store 디렉도리 구조 변경
store 디렉토리 하위에 store.js 로 store 객체 독립적으로 관리
store 디렉토리 하위에 carts, products 디렉토리를 도메인별로 나눠서 분리
carts, products 폴더 하위에 action, reducer로 파일 분리
