+++ src/index.jsx
@@ -1,19 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
-import { createStore, combineReducers, applyMiddleware } from 'redux';
-import thunk from 'redux-thunk';
 import App from 'App';
 import { worker } from 'mocks/browser';
-import { productsReducer } from 'store/products';
-import { cartsReducer } from 'store/carts';
-
-const rootReducer = combineReducers({
-  products: productsReducer,
-  carts: cartsReducer,
-});
-
-const store = createStore(rootReducer, applyMiddleware(thunk));
+import { store } from 'store/store';
 
 worker.start({
   serviceWorker: {
