+++ src/store/carts/action.js
@@ -1,19 +1,13 @@
 import axios from 'axios';
-import { SERVER_URL, PATH } from 'constants';
-import { ERROR_MESSAGE } from '../constants';
+import { SERVER_URL, PATH, ERROR_MESSAGE } from 'constants';
 
-const TYPE = {
+export const TYPE = {
   CARTS_LOAD: 'carts/load',
   CARTS_ADD: 'carts/add',
   CARTS_DELETE: 'carts/delete',
   CARTS_UPDATE: 'carts/update',
   CARTS_ERROR: 'carts/error',
   CARTS_CLEAN_ERROR: 'carts/clean',
-};
-
-const initialState = {
-  carts: [],
-  cartsError: null,
 };
 
 export const cartsActionCreators = {
@@ -71,41 +65,5 @@ export const updateCartQuantity = (product) => async (dispatch) => {
     dispatch(cartsActionCreators.updateCartQuantity(product));
   } catch (error) {
     dispatch(cartsActionCreators.error(ERROR_MESSAGE.UPDATE_CART_QUANTITY));
-  }
-};
-
-export const cartsReducer = (state = initialState, action) => {
-  switch (action.type) {
-    case TYPE.CARTS_LOAD:
-      return { ...state, carts: action.payload, cartsError: null };
-    case TYPE.CARTS_ADD:
-      return {
-        ...state,
-        carts: state.carts.concat(action.payload),
-        cartsError: null,
-      };
-    case TYPE.CARTS_DELETE:
-      return {
-        ...state,
-        carts: state.carts.filter((cart) => cart.id !== action.payload),
-        cartsError: null,
-      };
-    case TYPE.CARTS_UPDATE:
-      return {
-        ...state,
-        carts: state.carts.map((cart) => {
-          if (cart.id === action.payload.id) {
-            return action.payload;
-          }
-          return cart;
-        }),
-        cartsError: null,
-      };
-    case TYPE.CARTS_ERROR:
-      return { ...state, cartsError: action.payload };
-    case TYPE.CARTS_CLEAN_ERROR:
-      return { ...state, cartsError: null };
-    default:
-      return state;
   }
 };
